fix: pass the correct GDK Monitor
On dual monitor scaling configuration, seems that obtaining the Monitor
position using GDK yields the wrong results - both bars end up on the
same monitor.

We can use `monitor(idx: i32)` to obtain the correct one. However the
Rust bindings for GdkMonitor are not complete enough to let us call the
function that provides the connector name, therefore we need to query
smithay to get the vector and from the vector guess the index.

Fixes #1042.