feat(bluetooth): add options for popup height limit
* with #1186 gtk4 seemed to no longer set a min_height on the bluetoothbox, this adds a method to allow the box to grow until a limit of height is reached.

* combine scrollable in this new config parameter. and use the set_propagate_natural_height fn in combination with limiting max_content_height
If you are wondering why negative values and 0 are interpreted as No Scrollbar - TOML doesnt provide a possibilty to overwrite a default, so if we would rely on toml and have a default != null it would NOT be possible to ever deactivate the scrollbars :/
This is a flaw/design decision by toml. And ofc here it makes sense to have scrollbar by default.

* allow choosing between `pixels` and `devices` for Bluetooth module. and adjust docu